refactor: Move public script to TypeScript
